renamed asl::Endpoint to INetEndpoint to avoid name clashes
git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@4175 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ asl/net/TestSocket.cpp
@@ -210,7 +210,7 @@ void * TestSocket::TCPServerThread(void *arg) {
     TCPServer * theServer = reinterpret_cast<TCPServer*>(arg);
 
     char myInputBuffer[5];
-    int myResult = 0;
+    long myResult = 0;
 
     for (int i = 0; i<2; i++) {
         AC_DEBUG <<"TCPServerThread waiting to connect" << endl;
